src: move slot status file handling to status_file.c
This lived together with config file handling for quite a time now. But
since config file options become more and more, config_file.c grows and
status file handling, which never really fitted into its surrounding,
now gets a new and independent home in status_file.c to keep module size
and scope clearer.
