+++ src/main.c
@@ -20,6 +20,7 @@
 #include "rauc-installer-generated.h"
 #include "service.h"
 #include "signature.h"
+#include "status_file.h"
 #include "update_handler.h"
 #include "utils.h"
 #include "mark.h"
